+++ src/test/resources/fixtures/processing/blade/verbatim-and-comments-noop/expected.blade.php
@@ -1,0 +1,8 @@
+{{-- <div class="text-center p-4 flex bg-red-500 font-bold"></div> --}}
+
+@verbatim
+  <div class="text-center p-4 flex bg-red-500 font-bold"></div>
+  <div @class(['font-bold flex p-4' => active])></div>
+@endverbatim
+
+<div class="flex bg-red-500 p-4 text-center font-bold"></div>